requirements.txt
fastapi==0.109.2
google-cloud-bigquery==3.17.2
google-cloud-storage==2.14.0
google-generativeai==0.3.2
openai==1.12.0
python-dotenv==1.0.1
uvicorn==0.27.0.post1
